# coze/tool_openapi.yaml
openapi: 3.0.3
info:
  title: 数智链海受控业务工具
  version: 1.0.0
  description: |
    项目内部协议（v1），供 Coze 工作流以插件方式导入。
    所有接口需 Header `X-Context-Token`（由对话网关每轮签发，10 分钟有效）。
    工具接口验证 token；不信任模型传来的地区ID作为权限依据。
servers:
  - url: https://<项目后端域名>   # 部署时替换；Coze 云端无法访问 127.0.0.1
paths:
...
  /api/v1/tools/search_regional_knowledge:
    post:
      operationId: search_regional_knowledge
      summary: 地区授权资料检索（权限=服务端地区绑定，不信任模型自报地区）
      parameters:
        - in: header
          name: X-Context-Token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parameters]
              properties:
                parameters:
                  type: object
                  properties:
                    query: { type: string }
                    allowed_source_ids: { type: array, items: { type: string } }
      responses:
        '200':
          description: 片段（regional ref 带 source_id/scope/version；详情读取时再次检查权限）